InputView 구조체 생성
사용자가 입력력을 받는 함수 두개를 InputView 구조체를 만들어 분리하였습니다.
